backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
PyYAML==6.0.1
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
uuid==1.30
sqlalchemy==2.0.21
pydantic==2.4.2 